no cats

+++ src/config/ServerConfig.hpp
@@ -41,7 +41,7 @@ public:
 	const std::string& getListen() const; 
     const std::vector<std::string>& getServerNames() const;
     const std::map<std::string, std::string>& getErrorPages() const;
-    size_t getClientMaxBodySize() const;
+    long long getClientMaxBodySize() const;
     const std::string& getRoot() const;
     const std::string& getIndex() const;
     std::vector<LocationConfig>& getLocations();
@@ -53,7 +53,7 @@ public:
 	void setListen(const std::string& value); 
     void setServerNames(const std::vector<std::string>& value);
     void setErrorPages(const std::map<std::string, std::string>& value);
-    void setClientMaxBodySize(size_t value);
+    void setClientMaxBodySize(long long value);
     void setRoot(const std::string& value);
     void setIndex(const std::string& value);
     void setLocations(const std::vector<LocationConfig>& value);
@@ -66,7 +66,7 @@ private:
 	std::string listen; 
 	std::vector<std::string> serverNames;
     std::map<std::string, std::string> errorPages;
-    size_t clientMaxBodySize;
+    long long clientMaxBodySize;
     std::string root;
     std::string index;
     std::vector<LocationConfig> locations;
